+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # 0.3.0 (8 March 2026)
 
+- Invalidate `Item`s when the document changes
 - Remove `TomlError`
 - Bugfix inserting into inline table